@@ -1,7 +1,7 @@
 {
   "name": "ontimize-web-ngx",
   "homepage": "https://github.com/OntimizeWeb/ontimize-web-ngx#readme",
-  "version": "13.0.0",
+  "version": "14.0.0",
   "description": "Ontimize Web framework using Angular 8",
   "bugs": "https://github.com/OntimizeWeb/ontimize-web-ngx/issues",
   "author": "Imatia S.L.",
@@ -11,8 +11,8 @@
   },
   "keywords": [
     "angular",
-    "angular8",
-    "angular 8",
+    "angular14",
+    "angular 14",
     "framework",
     "web framework",
     "material",
@@ -34,18 +34,18 @@
     "test-watch": "karma start --singleRun=false --autoWatch=true"
   },
   "peerDependencies": {
-    "@angular/animations": "~13.4.0",
-    "@angular/cdk": "~13.3.9",
-    "@angular/common": "~13.4.0",
-    "@angular/compiler": "~13.4.0",
-    "@angular/core": "~13.4.0",
-    "@angular/flex-layout": "^13.0.0-beta.38",
-    "@angular/forms": "~13.4.0",
-    "@angular/material": "~13.3.9",
-    "@angular/material-moment-adapter": "~13.3.9",
-    "@angular/platform-browser": "~13.4.0",
-    "@angular/platform-browser-dynamic": "~13.4.0",
-    "@angular/router": "~13.4.0",
+    "@angular/animations": "^14.3.0",
+    "@angular/cdk": "^14.2.7",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/flex-layout": "^14.0.0-beta.40",
+    "@angular/forms": "^14.3.0",
+    "@angular/material": "^14.2.7",
+    "@angular/material-moment-adapter": "^14.2.7",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "rxjs": "~6.6.0",
     "zone.js": "~0.11.4",
     "google-libphonenumber": "^3.2.3"
