@@ -6,8 +6,11 @@ import { NgModule } from '@angular/core';
 import { MatPaginatorIntl } from '@angular/material/paginator';
 import { NgxMaterialTimepickerModule } from 'ngx-material-timepicker';
 
+import { PortalModule } from '@angular/cdk/portal';
 import { OSharedModule } from '../../shared/shared.module';
 import { OContextMenuModule } from '../contextmenu/o-context-menu.module';
+import { ODualListSelectorModule } from '../dual-list-selector/o-dual-list-selector.module';
+import { ODataToolbarModule } from '../o-data-toolbar/o-data-toolbar.module';
 import { OTableColumnCalculatedComponent } from './column/calculated/o-table-column-calculated.component';
 import { O_TABLE_CELL_EDITORS } from './column/cell-editor/cell-editor';
 import { O_TABLE_CELL_RENDERERS } from './column/cell-renderer/cell-renderer';
@@ -16,19 +19,15 @@ import { OTableContextMenuComponent } from './extensions/contextmenu/o-table-con
 import { O_TABLE_DIALOGS } from './extensions/dialog/o-table-dialog-components';
 import { OTableExportButtonComponent } from './extensions/export-button/o-table-export-button.component';
 import { OTableExportButtonService } from './extensions/export-button/o-table-export-button.service';
-import { OTableColumnAggregateComponent } from './extensions/footer/aggregate/o-table-column-aggregate.component';
 import { OTableExpandedFooterDirective } from './extensions/footer/expanded/o-table-expanded-footer.directive';
 import { O_TABLE_FOOTER_COMPONENTS } from './extensions/footer/o-table-footer-components';
 import { OTableMatPaginatorIntl } from './extensions/footer/paginator/o-table-mat-paginator-intl';
 import { O_TABLE_HEADER_COMPONENTS, O_TABLE_HEADER_COMPONENTS_EXPORTED } from './extensions/header/o-table-header-components';
 import { OTableRowClassPipe } from './extensions/pipes/o-table-row-class.pipe';
 import { OTableRowDirective } from './extensions/row/o-table-row.directive';
+import { OTableRowExpandableComponent } from './extensions/row/table-row-expandable/o-table-row-expandable.component';
 import { OMatSortModule } from './extensions/sort/o-mat-sort-module';
 import { OTableComponent } from './o-table.component';
-import { OTableRowExpandableComponent } from './extensions/row/table-row-expandable/o-table-row-expandable.component';
-import { PortalModule } from '@angular/cdk/portal';
-import { ODualListSelectorModule } from '../dual-list-selector/o-dual-list-selector.module';
-import { ODataToolbarModule } from '../o-data-toolbar/o-data-toolbar.module';
 
 @NgModule({
     declarations: [
